% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel.R
\name{iControl}
\alias{iControl}
\title{Control parameters for RFT based analyses}
\usage{
iControl(cf = 0.05, scr = TRUE, sar = 64, n = 1, iso = TRUE, os = 3,
  rft = TRUE)
}
\arguments{
\item{cf}{Critical F-threshold for selecting voxels over which the non-sphericity is estimated (default = \code{0.001}).}

\item{scr}{Logical. scale residuals? (default = \code{TRUE}).}

\item{sar}{Number of residual images to sample for estimating the FWHM (default = \code{64}).}

\item{n}{images in conjunction (default = \code{1}).}

\item{iso}{logical. should images be assumed to be isotropic? (default = \code{TRUE}).}

\item{os}{offset weighting for iteratively reweighted least squares (default = \code{3}).}

\item{rft}{logical. should voxels be estimated in resel space for random field theory analysis (default = \code{TRUE}).}
}
\value{

}
\description{
Auxillary function for controlling \code{\link{iModel}} fitting.
}

