% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iData.R
\docType{class}
\name{iData-class}
\alias{iData}
\alias{iData-class}
\title{Class iData}
\arguments{
\item{.Object}{input object to convert}

\item{x}{Either a list of iGroups or a single iGroup object}

\item{bool}{A vector of TRUE/FALSE values with length equal to the number of
rows in the demographics data frame and number of TRUE values equal to the 
number of rows in the image matrix.}

\item{demog}{Demographic data.frame corresponding to iGroups.}
}
\description{
Object for representing multiple image modalities and demographic data
}
\section{Slots}{

\describe{
\item{\code{iList}}{List of iGroup objects.}

\item{\code{demog}}{Demographic information.}

\item{\code{index}}{Index that coordinates iGroup rows with demographic rows.}
}}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{iGroup}}, \code{\link{iData-methods}}
}

