% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iClass.R
\docType{class}
\name{iGroup-class}
\alias{iGroup}
\alias{iGroup-class}
\title{Class iGroup}
\arguments{
\item{.Object}{Inpute object to convert.}

\item{iMatrix}{Image by voxel matrix.}

\item{name}{Name for the iGroup object (used for reference in \code{\link{iData}}
and \code{\link{iFormula}})}

\item{mask}{Mask with number of non-zero entries defining the matrix columns.}

\item{modality}{Image modality that iGroup will represent.}

\item{rowslist}{Row of iMatrix constituting block/partitions.}

\item{HParam}{Cut-off period in seconds.}

\item{RT}{Observation interval in seconds.}

\item{checkMask}{Logical ensure mask only represents active voxels (default
= \code{TRUE}).}

\item{filename}{Optional filename to save iGroup object (default = 
\code{tempfile(fileext = ".h5")}).}
}
\description{
Object for representing single image modality and preserving active memory.
}
\section{Slots}{

\describe{
\item{\code{name}}{Name of the iGroup.}

\item{\code{file}}{h5file connection.}

\item{\code{iMatrix}}{h5file pointer to matrix.}

\item{\code{location}}{h5file location.}

\item{\code{modality}}{Image modality represented.}

\item{\code{mask}}{Mask with number of non-zero entries defining the matrix columns.}

\item{\code{K}}{Filter information.}

\item{\code{dim}}{Dimensions of iMatrix.}
}}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{iGroup-methods}}, \code{\link{iData-class}}
}

