% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iClass.R
\docType{methods}
\name{iGroup-methods}
\alias{[,iGroup,ANY,ANY,ANY-method}
\alias{dim,iGroup-method}
\alias{iGroup-methods}
\alias{iGroupMask}
\alias{iGroupRead}
\alias{iGroupWrite}
\alias{names,iGroup-method}
\alias{names<-,iGroup-method}
\title{iGroup Methods}
\usage{
\S4method{dim}{iGroup}(x)

\S4method{names}{iGroup}(x)

\S4method{names}{iGroup}(x) <- value

\S4method{[}{iGroup,ANY,ANY,ANY}(x, i)

iGroupMask(x, mask)

iGroupRead(filename)

iGroupWrite(x, filename)
}
\arguments{
\item{x, object}{Object of class iGroup.}

\item{filename}{h5 file to save iGroup object to}
}
\description{
iGroup Methods
}
\details{
\strong{dim} Retrieve dimensions of iGroup's iMatrix slot.

\strong{names} Retrieve name of iGroup object.

\strong{names<-} Set name of iGroup.

\strong{iGroup[i]} Set name of iGroup.

\strong{iGroupMask} mask iGroup object to specific region.

\strong{iGroupRead} Read/load iGroup from h5 file.

\strong{iGroupWrite} Write/save iGroup to h5 file.
}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{iGroup-class}}
}

