% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel.R
\name{ilm}
\alias{ilm}
\title{Fit Images to Linear Models}
\usage{
ilm(formula, iData, weights, na.action = "na.omit", control = list(),
  verbose = TRUE)
}
\arguments{
\item{formula}{Object of class \code{formula} (or one that can be coerced to that class): a symbolic description of the model to be fitted. The details of model specification are given under ‘Details’.}

\item{iData}{Object of class \code{\link{iData}} containing data represented in the provided formula.}

\item{weights}{An optional vector of ‘prior weights’ to be used in the fitting process. Should be NULL or a numeric vector.}

\item{na.action}{If \code{"na.omit"} then all images with NA values in 
corresponding variables are omitted. If a function is specified}

\item{control}{A list of parameters for controlling the fitting process. See \code{\link{iControl}} for details.}

\item{verbose}{Enables verbose output. (default = \code{TRUE}).}
}
\description{
Fit Images to Linear Models
}

