% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel.R
\docType{class}
\name{iModel-class}
\alias{iModel}
\alias{iModel-class}
\alias{iglModel}
\alias{iglModel-class}
\alias{ilModel}
\alias{ilModel-class}
\title{Class iModel}
\description{
Object containing a fitted statistical model for image data.
}
\section{Slots}{

\describe{
\item{\code{iData}}{iData reference class object for fitted model.}

\item{\code{Y}}{Pointer to image matrix stored as h5 dataset.}

\item{\code{X}}{Information about the design:
\itemize{
 \item{X} {Design matrix.}
 \item{K} {Filter information specific to each image.}
 \item{W} {weights}
 \item{KWX} {Weighted design matrix information.}
 \item{pKX} {Pseudoinverse of X.}
 \item{V} {Non-sphericity matrix.}
 \item{betaCov} {Covariance of correlation coefficients.}
 \item{trRV} {Trace of RV.}
 \item{trRVRV} {Trace of RVRV.}
 \item{rdf} {Residual degrees of freedom.}
}}

\item{\code{beta}}{Pointer to beta matrix stored as h5 dataset.}

\item{\code{res}}{Pointer to residual matrix stored as h5 dataset.}

\item{\code{mrss}}{Pointer to mean residual sum of squares matrix stored as h5 dataset.}

\item{\code{xVi}}{Information about intrinsic temporal non-sphericity:
\itemize{
 \item{Vi} {List of non-sphericity components.}
 \item{V} {Non-sphericity matrix (Cov(e) = sigma^2*V)}
 \item{h} {Hyperparameters.}
 \item{Cy} {Covariance of response matrix.}
}}

\item{\code{dims}}{Model dimensions:
\itemize{
 \item{npred} {Number of predictors.}
 \item{nimg} {Number of images.}
 \item{nvox} {Number of voxels.}
 \item{fwhm} {full-width at half-maxima}
 \item{resels} {Resolution elements.}
 \item{rpvImage} {Resels per voxel image.}
}}

\item{\code{C}}{list of contrasts:
\itemize{
 \item{name} {Name of the contrast.}
 \item{c} {Contrast weights.}
 \item{X1} {Reamaining design space (orthogonal to X0)}
 \item{X0} {Reduced design matrix.}
 \item{iX0} {Indicates how contrat was specified.}
 \item{dims} {Dimensions for contrast}
 \itemize{
   \item{trRV} {Trace of RV.}
   \item{trRVRV} {Trace of RVRV.}
   \item{idf} {Degrees of interest.}
 }
 \item{fieldType} {Type of statyistical field being fitted.}
 \item{contrastImage} {Object of class antsImage representing contrast.}
 \item{clusterImage} {Object of class antsImage representing the thresholded contrastImage.}
 \item{results} {Either a list or data.frame containing the results of a given contrast.}
 \item{sthresh} {Statistical threshold.}
 \item{cthresh} {Cluster threshold.}
}}

\item{\code{control}}{Control values for fitting the model (see \code{\link{iControl}}).}
}}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{ilm}}, \code{\link{iglm}}
}

