% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iContrast.R
\docType{methods}
\name{iModel-contrasts}
\alias{anova,iModel-method}
\alias{iModel-contrasts}
\alias{summary,iModel-method}
\title{iModel Contrasts}
\usage{
\S4method{anova}{iModel}(object, contrastMatrix, cthresh = 150,
  threshType = "cFDR", pval = 0.05, pp = 0.001, verbose = TRUE)

\S4method{summary}{iModel}(object, contrastMatrix, cthresh = 150,
  threshType = "pRFT", pval = 0.05, pp = 0.001, verbose = TRUE)
}
\arguments{
\item{object}{Object of class iModel.}

\item{contrastMatrix}{Matrix of weighted contrasts in which each row 
corresponds to an individual contrast and each column of each row
corresponds to a predictor's/variable's specified weights.}

\item{cthresh}{Minimum desired cluster size (default = \code{150})}

\item{threshType}{A numeric value to threshold the statistical field or a character of the following methods:
\itemize{
\item{cRFT} {Computes a threshold per expected cluster level probability.}
\item{pRFT} {Uses the mask and pval calculates the minimum statistical threshold.}
\item{cFDR} {Uses an uncorrected threshold at the alpha level and then computes and FDR threshold based on cluster maxima.}
\item{pFDR} {Computes the fdr threshold for the entire field of voxels.}
}}

\item{pval}{The p-value for estimating the threshold (default = \code{0.05}).}

\item{pp}{The primary (initial) p-value for thresholding (only used for FDR methods; default = \code{0.001}).}

\item{verbose}{Enables verbose output.}
}
\description{
iModel Contrasts

\strong{anova} Compute f-statistic or t-statistic contrast for iModel object.

\strong{summary} Compute f-statistic or t-statistic contrast for iModel object.
}

