% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iModel.R
\docType{methods}
\name{iModelSolve}
\alias{coef.iModel}
\alias{fitted.iModel}
\alias{getImages}
\alias{iModel-methods}
\alias{iModelRFT}
\alias{iModelRead}
\alias{iModelSolve}
\alias{iModelUpdate}
\alias{iModelWrite}
\alias{model.matrix.iModel}
\alias{report}
\alias{resid.iModel}
\title{iModel Methods}
\usage{
iModelSolve(object, verbose = TRUE)

iModelUpdate(object, ...)

iModelRFT(object, verbose = TRUE)

getImages(object, contrast, rpv = FALSE, mask = FALSE)

\method{coef}{iModel}(object)

\method{fitted}{iModel}(object)

\method{resid}{iModel}(object)

iModelRead(filename)

iModelWrite(object, filename)

iModelSolve(object, verbose = TRUE)

iModelUpdate(object, ...)

\method{model.matrix}{iModel}(object)

report(object, dirname, surfimg)
}
\arguments{
\item{object}{Object of class iModel.}

\item{verbose}{Enables verbose output. (default = \code{TRUE}).}

\item{...}{Additional named arguments passed to \code{iModelUpdate}.}

\item{contrast}{Name of contrast.}

\item{rpv}{Return resels per voxel image?}

\item{mask}{Return mask used for iModel?}

\item{filename}{h5 file to save iModel to.}

\item{dirname}{Directory to place \code{report} output.}

\item{iData_dirname}{Directory for iData component of iModel.}
}
\description{
V
}
\details{
\strong{iModelSolve} Solve already initialized iModel for
 coefficients, residuals, and mean residual sum of squares.

\strong{iModelUpdate} Primarily used to update X slot after
 optimization.

\strong{iModelRFT} Solve for images in resel space.

\strong{getImages} Returns antsImages of specified contrasts within model iModel.

\strong{coef} Retrieve fitted coefficients from iModel object.

\strong{fitted} Retrieve fitted values from iModel object.

\strong{resid} Retrieve residuals from iModel object.

\strong{iModelRead} Read/load iModel object.

\strong{iModelWrite} Read/load iModel object.

\strong{iModelSolve} Solve already initialized iModel for
 coefficients, residuals, and mean residual sum of squares.

\strong{iModelUpdate} Primarily used to update X slot after optimization

\strong{model.matrix} Retrieve design matrix from iModel object.

\strong{report} Creates a .pdf and .html report of for iModel
 objects (requires package rmarkdown).
}
\author{
Zachary P. Christensen
}

