% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rftResults.R
\name{rftResults}
\alias{rftResults}
\title{Produce Random Field Theory Statistical Results}
\usage{
rftResults(x, resels, fwhm, df, fieldType, rpvImage = NULL, k = 1,
  threshType = "pRFT", pval = 0.05, pp = 0.001, n = 1,
  verbose = FALSE)
}
\arguments{
\item{x}{statistical field image of class antsImage}

\item{resels}{resel values for the mask}

\item{fwhm}{full width at half maxima}

\item{df}{degrees of freedom expressed as df = c(degrees of interest, degrees of error)}

\item{rpvImage}{resels per voxel image}

\item{k}{minimum desired cluster size (default = 1)}

\item{threshType}{a numeric value to threshold the statistical field or a character of the following methods:
\itemize{
\item{cRFT:} {computes a threshold per expected cluster level probability }
\item{pRFT:} {uses the mask and pval calculates the minimum statistical threshold}
\item{cFDR:} {uses an uncorrected threshold at the alpha level and then computes and FDR threshold based on cluster maxima}
\item{pFDR:} {computes the fdr threshold for the entire field of voxels}
}}

\item{pval}{the p-value for estimating the threshold (default = .05)}

\item{pp}{the primary (initial) p-value for thresholding (only used for FDR methods; default = .001)}

\item{n}{number of images in conjunction}

\item{verbose}{enables verbose output}

\item{fieldType:}{\itemize{
\item{"T"}{T-field} 
\item{"F"}{F-field} 
\item{"X"}{Chi-square field"} 
\item{"Z"}{Gaussian field}
}}
}
\value{
Outputs a statistical value to be used for threshold a statistical field image

\item{setLevel:} {set-level statistics and number of clusters}
\item{clusterLevel:} {cluster-level statistics and descriptors}
\item{peakLevel:} {peak-level statistics and descriptor"}
\item{clusterImage:} {image of labeled clusters}
\item{threshold:} {the threshold used}
}
\description{
Returns RFT based statistical results for a single statistical image
}
\details{
\code{rftPval} is used to compute all family-wise error (FWE) corrected
statistics while \code{\link{p.adjust}} is used to compute all false-discovery rate
based statistics. All statistics herein involve implementation of random
field theory (RFT) to some extent.

Both cluster-level and peak-level statistics are described by the uncorrected
p-value along with the FDR and FWE corrected p-values for each cluster.
Peak-level statistics are described by the maximum statistical value in each
cluster. The ClusterStats table also contains
coordinates for each cluster and the number of voxels therein. By default
\code{\link{threshType = "pRFT"}} and pval=.05. Alternatively, the user may use a
specific numeric value for thresholding the statistical field.
\code{\link{statFieldThresh}} more fully describes using appropriate thresholds
for statistical fields and how \code{pp} plays a role in FDR thresholding.
}
\examples{
\dontrun{
mnit1 <- antsImageRead(getANTsRData('mni'))
mask <- getMask(mnit1)
ilist <- list()
for (i in 1:10) {
 ilist <- lappend(ilist, antsImageClone(mnit1) * rnorm(1))
}
response <- rnorm(10)
imat <- imageListToMatrix(ilist, mask)
residuals <- matrix(nrow = nrow(imat), ncol = ncol(imat))
tvals <- matrix(nrow = nrow(imat), ncol = ncol(imat))
for (i in 1:ncol(imat)) {
 fit <- lm(response ~ imat[, i])
 tvals <- coefficients(fit)[2]
 residuals[, i] <- residuals(fit)
}
myfwhm <- estSmooth(residuals, mask, fit$df.residual)
res <- resels(mask, myfwhm$fwhm)
timg <- makeImage(mask, tvals)

# threshold to create peak values with p-value of .05 (default)
results1 <- rftResults(timg, res, myfwhm$fwhm, df, fieldType = "T",
                       threshType = "pRFT", pval = .05)

# threshold to create clusters with p-value of .05
results2 <- rftResults(timg, res, myfwhm$fwhm, df, fieldType = "T",
                       threshType = "cRFT", pval = .05)

# initial threshold at p-value of .001 followed by peak FDR threshold at
# p-value of .05
results3 <- rftResults(timg, res, myfwhm$fwhm, df, fieldType = "T",
                       threshType = "pFDR", pval = .05, pp=.01)

# initial threshold at p-value of .001 followed by cluster FDR threshold at
# p-value of .05
results4 <- rftResults(timg, res, myfwhm$fwhm, df, fieldType = "T",
                       threshType = "cFDR", pval = .05, pp = .01)

# correcting for non-isotropic
results5 <- rftResults(timg, res, myfwhm$fwhm, df, fieldType = "T",
                       fwhm$rpvImage)

}
}
\author{
Zachary P. Christensen
}
\references{
Chumbley J., (2010) Topological FDR for neuroimaging

Friston K.J., (1996) Detecting Activations in PET and fMRI: Levels of Inference and Power

Worsley K.J., (1992) A Three-Dimensional Statistical Analysis for CBF Activation Studies in Human Brain.
}
\keyword{\code{\link{rftPval}}}

