% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iClass.R
\docType{class}
\name{iData-class}
\alias{iData}
\alias{iData-class}
\title{Class iData}
\arguments{
\item{.Object}{input object to convert}

\item{x}{either a list of iGroups or a single iGroup object}

\item{bool}{a vector of TRUE/FALSE values with length equal to the number of
rows in the demographics data frame and number of TRUE values equal to the 
number of rows in the image matrix.}

\item{demog}{demographic data.frame corresponding to iGroups}
}
\description{
Object for representing multiple image modalities and demographic data
}
\section{Slots}{

\describe{
\item{\code{iList}}{list of iGroup objects}

\item{\code{demog}}{demographic information}

\item{\code{index}}{index that coordinates iGroup rows with demographic rows}
}}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{iGroup}}, \code{\link{iData-methods}}
}

