% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iGroup.R, R/iGroup_methods.R
\docType{methods}
\name{iGroup-methods}
\alias{[,iGroup,ANY,ANY-method}
\alias{colMeans,iGroup-method}
\alias{colSums,iGroup-method}
\alias{dim,iGroup-method}
\alias{iGroup-methods}
\alias{iGroupMask}
\alias{iGroupRead}
\alias{iGroupWrite}
\alias{max,iGroup-method}
\alias{mean.iGroup}
\alias{min,iGroup-method}
\alias{names,iGroup-method}
\alias{names<-,iGroup-method}
\alias{range,iGroup-method}
\alias{rowMeans,iGroup-method}
\alias{rowSums,iGroup-method}
\alias{sd,iGroup-method}
\alias{sum,iGroup-method}
\title{iGroup Methods}
\usage{
\S4method{dim}{iGroup}(x)

\S4method{names}{iGroup}(x)

\S4method{names}{iGroup}(x) <- value

\S4method{[}{iGroup,ANY,ANY}(x, i)

iGroupMask(x, mask)

iGroupRead(filename)

iGroupWrite(x, filename)

\S4method{max}{iGroup}(x)

\S4method{min}{iGroup}(x)

\S4method{range}{iGroup}(x)

\S4method{rowSums}{iGroup}(x)

\S4method{rowMeans}{iGroup}(x)

\S4method{colSums}{iGroup}(x)

\S4method{colMeans}{iGroup}(x)

\S4method{sum}{iGroup}(x)

mean.iGroup(x)

\S4method{sd}{iGroup}(x)
}
\arguments{
\item{x, object}{Object of class iGroup.}

\item{filename}{h5 file to save iGroup object to}
}
\description{
iGroup Methods
}
\details{
\strong{dim} Retrieve dimensions of iGroup's iMatrix slot.

\strong{names} Retrieve name of iGroup object.

\strong{names<-} Set name of iGroup.

\strong{iGroup[i]} Set name of iGroup.

\strong{iGroupMask} mask iGroup object to specific region.

\strong{iGroupRead} Read/load iGroup from h5 file.

\strong{iGroupWrite} Write/save iGroup to h5 file.

\strong{max}

\strong{min}

\strong{range}

\strong{rowSums}

\strong{rowMeans}

\strong{colSums}

\strong{colMeans}

\strong{sum}

\strong{mean}

\strong{sd}
}
\author{
Zachary P. Christensen
}
\seealso{
\code{\link{iGroup-class}}
}

